fix: update FailedTestsTab
This updates the CodecovCLI page in the FailedTestsTab to use the
FrameworkTabsCard. This also reorders the steps on the CodecovCLI page
such that the JUnit XML generation step comes first for consistency
with the other page.
