+++ src/pages/RepoPage/FailedTestsTab/FrameworkTabsCard/FrameworkTabs/index.ts
