+++ src/pages/RepoPage/FailedTestsTab/FrameworkTabsCard/index.ts
@@ -1,0 +1,1 @@
+export { FrameworkTabsCard } from './FrameworkTabsCard'
